Cache-bust asset files based on version
Add ?version=x.x.x to the end of the js and css files, so that, when a
user upgrades to a newer version of Video Link, they won't get cached
assets files that either cause issues or don't pull in fixed issues.

The version is pulled automatically from the version of Video Link defined
in the fieldtype file.