Store and output title, thumbnail
On the Publish field, keep track of both the title and the thumbnail, in
addition to the URL, and send them back to the server. This allows us to
output the thumbnail or title when displaying the field, without having to
query the YouTube and Vimeo endpoints at display time.

Fixes #3